% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{aic_tab}
\alias{aic_tab}
\title{radish model selection}
\usage{
XXX
}
\arguments{
\item{mod_list}{List containing fitted `radish` objects}

\item{AICc}{Use second order AIC in ranking models (Default = FALSE). See Details}

\item{BIC}{Use BIC in ranking models (Default = FALSE). See Details}

\item{mod_names}{Optional. Specify names for fitted model objects. By default, the right hand side of the fitted `radish` model will be used as the model name.}

\item{verbose}{(Default = FALSE) Should the table be pronted to the console}

\item{...}{Not used}
}
\value{
Data frame with AIC summary table for provided models
}
\description{
Function to create AIC table of fitted models
}
\examples{
## Not Run:
## ** TO BE COMPLETED ** ##  

## End (Not run)
}
\author{
Bill Peterman
}
