% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin2cont_function.R
\name{bin2cont}
\alias{bin2cont}
\title{Make a binary raster continuous}
\usage{
bin2cont(bin_rast, window = 3)
}
\arguments{
\item{bin_rast}{Binary (1/0) raster}

\item{window}{The size of the moving window to calculate mean habitat within (Default = 3)}
}
\description{
Function to create a continuous surface from a binary surface
}
\author{
Bill Peterman
}
