% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdpop_from_R_function.R
\name{cdpop}
\alias{cdpop}
\title{Internal CDPOP function}
\usage{
cdpop(
  CDPOP.py,
  sim_name = "output_",
  pts,
  sim_dir,
  resist_rast,
  resist_mat = NULL,
  agefilename = NULL,
  mcruns = 1,
  looptime = 400,
  output_years = 50,
  gridformat = "genepop",
  cdclimgentime = 0,
  matemoveno = 2,
  matemoveparA = 0,
  matemoveparB = 0,
  matemoveparC = 0,
  matemovethresh = "max",
  output_matedistance = "N",
  sexans = "Y",
  Freplace = "N",
  Mreplace = "N",
  philopatry = "N",
  multiple_paternity = "N",
  selfans = "N",
  Fdispmoveno = NULL,
  FdispmoveparA = 0,
  FdispmoveparB = 0,
  FdispmoveparC = 0,
  Fdispmovethresh = NULL,
  Mdispmoveno = NULL,
  MdispmoveparA = 0,
  MdispmoveparB = 0,
  MdispmoveparC = 0,
  Mdispmovethresh = NULL,
  offno = 2,
  MeanFecundity = 5,
  Femalepercent = 50,
  EqualsexratioBirth = "N",
  TwinningPercent = 0,
  popModel = "exp",
  r = 1,
  K_env = length(pts),
  subpopmortperc = 0,
  muterate = 5e-04,
  mutationtype = "random",
  loci = 1000,
  intgenesans = "random",
  allefreqfilename = "N",
  alleles = 2,
  mtdna = "N",
  startGenes = 0,
  cdevolveans = "N",
  startSelection = 0,
  betaFile_selection = "N",
  epistasis = "N",
  epigeneans = "N",
  startEpigene = 0,
  betaFile_epigene = "N",
  cdinfect = "N",
  transmissionprob = 0,
  deme_size = 1,
  n_demes = NULL,
  python = NULL
)
}
\arguments{
\item{CDPOP.py}{Full path to the CDPOP.py file}

\item{sim_name}{Name for simulation results. Defaults to 'output'}

\item{pts}{Spatial points object}

\item{sim_dir}{Directory where simulation results will be written}

\item{resist_rast}{Resistance surface}

\item{agefilename}{Path to age file. Default will create and use a non-overlapping generations file.}

\item{mcruns}{Default = 1}

\item{looptime}{Default = 401; Number generations to conduct simulation}

\item{output_years}{Default = 50; Interval to write out simulation results}

\item{gridformat}{Default = 'genepop'; c('genepop', 'genalex', 'structure', 'cdpop')}

\item{cdclimgentime}{Default = 0. To initiate the CDClimate module, this is the generation/year that the next effective distance matrix will be read in at. You can specify multiple generations by separating each generation to read in the next cost distance matrix by ‘|’.  Then in the following surface columns, a separate file can be given for each generation.}

\item{matemoveno}{Default = 2; Uses Inverse Square (1 / (Cost Distance^2)). This function gets rescaled to min and threshold of the inverse square cost distance.}

\item{matemoveparA}{Not used with inverse square movement}

\item{matemoveparB}{Not used with inverse square movement}

\item{matemoveparC}{Not used with inverse square movement}

\item{matemovethresh}{Default = 'max'; The maximum movement is the maximum resistance distance}

\item{output_matedistance}{Does mate movement distance differ (Default = 'N')}

\item{sexans}{Default = 'N'; No selfing}

\item{Freplace}{Default = 'N'; Females mate without replacement}

\item{Mreplace}{Default = 'N'; Males mate without replacement}

\item{philopatry}{Default = 'N';}

\item{multiple_paternity}{Default = 'N'; No philopatry of Males or Females}

\item{selfans}{Default = 'N'; No selfing}

\item{Fdispmoveno}{Default = NULL. Will be set equal to matemoveno}

\item{FdispmoveparA}{Not used with inverse square movement}

\item{FdispmoveparB}{Not used with inverse square movement}

\item{FdispmoveparC}{Not used with inverse square movement}

\item{Fdispmovethresh}{Default = NULL. Will be set to matemovethresh}

\item{Mdispmoveno}{Default = NULL. Will be set equal to matemoveno}

\item{MdispmoveparA}{Not used with inverse square movement}

\item{MdispmoveparB}{Not used with inverse square movement}

\item{MdispmoveparC}{Not used with inverse square movement}

\item{Mdispmovethresh}{Default = NULL. Will be set to matemovethresh}

\item{offno}{Default = 2; Poisson draw around ‘mean fecundity’}

\item{MeanFecundity}{Default = 5; Specifies mean fecundity in age variable file.}

\item{Femalepercent}{Default = 50}

\item{EqualsexratioBirth}{Default = 'N'}

\item{TwinningPercent}{Default = 0}

\item{popModel}{Default = 'exp'}

\item{r}{Population growth rate. No applicable when using exponential growth rate}

\item{K_env}{Equal to the number of individuals simulated}

\item{subpopmortperc}{Default = 0|0|0|0; Not using subpopulation features}

\item{muterate}{Default = 0.0005}

\item{mutationtype}{Default = 'forward'}

\item{loci}{Default = 1000; For simulating SNP-like markers}

\item{intgenesans}{Default = 'random'; Random initiation of alleles}

\item{allefreqfilename}{Default = 'N'}

\item{alleles}{Default = 2; For simulating SNP-like markers}

\item{mtdna}{Default = 'N'}

\item{startGenes}{Default = 0;}

\item{cdevolveans}{Default = 'N'; No loci are under selection}

\item{startSelection}{Default = 0; No selection}

\item{betaFile_selection}{Default = 'N'; No selection}

\item{epistasis}{Default = 'N'; No epigenetics}

\item{epigeneans}{Default = 'N'; No epigenetics}

\item{startEpigene}{Default = 0; No epigenetics}

\item{betaFile_epigene}{Default = 'N'; No epigenetics}

\item{cdinfect}{Default = 'N'; No epigenetics}

\item{transmissionprob}{Default = 0; No epigenetics}

\item{deme_size}{Number of individuals within each deme}

\item{n_demes}{Number of demes on the landscape}

\item{python}{Optional. Only needed if Python 2.7 is not in your system path. If needed, specify the full path to the Python 2.7x program file.}
}
\description{
Function to run CDPOP from R
}
\author{
Bill Peterman
}
\keyword{internal}
