% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlated_raster_function.R
\name{corr_rast}
\alias{corr_rast}
\title{Generate correlated rater surfaces}
\usage{
corr_rast(corr, FUN, ...)
}
\arguments{
\item{corr}{Desired correlation level between fbm surfaces}

\item{FUN}{NLMR function to create random surface}

\item{...}{Specify all necessary parameters of the function `FUN`}
}
\description{
Function to create correlated surfaces
}
\examples{
## Not Run:
library(raster)
cr <- corr_rast(corr = 0.5,
                FUN = NLMR::nlm_fbm,
                ncol = 50,
                nrow = 50,
                fract_dim = 0.5,
                user_seed = 555) 
plot(cr) 

## End (Not run)
}
\author{
Bill Peterman
}
