% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radishCV_selection.R
\name{cv_model_selection}
\alias{cv_model_selection}
\title{Conduct model selection on cross validation models}
\usage{
XXX
}
\arguments{
\item{cv_list}{List containing fitted cross validation objects}

\item{cv_names}{Optional. Specify names for fitted cv objects. By default, the right hand side of the fitted `radish` model will be used as the name.}

\item{aic}{If the full radish model was fit during cross validation, an AIC table can also be created (Default = FALSE)}

\item{...}{Not used}
}
\value{
Data frame or named list with loglikelihood table and AIC summary table for provided models
}
\description{
Function to create model selection tables from fitted cross validation models
}
\examples{
## Not Run:
## ** TO BE COMPLETED ** ##  

## End (Not run)
}
\author{
Bill Peterman
}
