% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlated_fbm_function.R
\name{fbm_corr}
\alias{fbm_corr}
\title{Generate correlated FBM surfaces}
\usage{
fbm_corr(corr, dim, H = 0.5, user_seed = NA)
}
\arguments{
\item{corr}{Desired correlation level between fbm surfaces}

\item{dim}{Dimension of simulated raster (Default = 400)}

\item{H}{Fractal dimension (Hurst parameter, H) controlling clustering in Fractal Brownian Motion landscape simulation model (Default = 0.5)}

\item{user_seed}{Seed can be set for reproducibility (Default = NA)}
}
\description{
Function to create correlated Fractal Brownian Motion surfaces
}
\author{
Bill Peterman
}
