% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlated_gaus_function.R
\name{gaus_corr}
\alias{gaus_corr}
\title{Generate correlated Gaussian surfaces}
\usage{
gaus_corr(corr, dim, autocorr_range = 5, mag_var = 25, nug = 5, user_seed = NA)
}
\arguments{
\item{corr}{Desired correlation level between fbm surfaces}

\item{dim}{Dimension of simulated raster}

\item{autocorr_range}{Maximum range (raster units) of spatial autocorrelation. (Default = 5)}

\item{mag_var}{Magnitude of variation over the entire landscape. (Default = 25)}

\item{nug}{Magnitude of variation in the scale of autocorr_range, smaller values lead to more homogeneous landscapes.(Default = 5)}

\item{user_seed}{Seed can be set for reproducibility (Default = NA)}
}
\description{
Function to create correlated Gaussian surfaces
}
\author{
Bill Peterman
}
