% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gausDGS.R
\name{gausDGS}
\alias{gausDGS}
\title{Gaussian surface}
\usage{
gausDGS(
  dim = 300,
  autocorr_range = NULL,
  mag_var = 5,
  nug = 1,
  user_seed = NULL
)
}
\arguments{
\item{dim}{Number rows and columns in raster (Default 300)}

\item{autocorr_range}{Maximum range (raster units) of spatial autocorrelation (Default will be 3 percent of raster 'dim')}

\item{mag_var}{Magnitude of variation over the entire landscape (Default 5)}

\item{nug}{Magnitude of variation in the scale of autocorr_range, smaller values lead to more homogeneous landscapes (Default 1)}

\item{user_seed}{Set random seed for the simulation}
}
\description{
Function to create Gaussian random field raster surface for use with DGS simulation
}
\author{
Bill Peterman
}
