% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{Retrieve data}
\usage{
XXX
}
\arguments{
\item{dir}{Full path to directory where simulation results saved}

\item{...}{Not used}
}
\value{
Data frame or named list with loglikelihood table and AIC summary table for provided models
}
\description{
Function to retrieve CDPOP simulation files for `radish` analysis
}
\examples{
## Not Run:
## ** TO BE COMPLETED ** ##  

## End (Not run)

}
\author{
Bill Peterman
}
