% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_thresh_function.R
\name{habitat_threshold}
\alias{habitat_threshold}
\title{Specify habitat threshold}
\usage{
habitat_threshold(hab_rast, p)
}
\arguments{
\item{hab_rast}{The simulated habitat surface}

\item{p}{The proportion of the landscape that should be habitat (0-1)}
}
\description{
Function to create binary raster from continuous habitat surface
}
\author{
Bill Peterman
}
