% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_funcs.R
\name{pca_dist}
\alias{pca_dist}
\title{PCA distance}
\usage{
pca_dist(gi, n_axes = 64, scale = T)
}
\arguments{
\item{gi}{`genind` object from adegenet}

\item{n_axes}{Number of principle component axes to retain and calculate distance between}

\item{scale}{(Default = TRUE) Calculated distance will be rescaled 0-1}
}
\description{
Calculate Euclidean distance between principle component axes
}
\author{
Bill Peterman
}
