% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cdpop.R
\name{plot_cdpop}
\alias{plot_cdpop}
\title{Plot simulations}
\usage{
XXX
}
\arguments{
\item{params}{Object generated from `get_params` OR results from `cdpop_sim`}

\item{...}{Not used}
}
\value{
4-panel plot of two genetic distances X resistance distance and Euclidean distance
}
\description{
Function to plot genetic distances generated from CDPOP simulations
}
\examples{
## Not Run:
## plot_cdpop(params) ##  

## End (Not run)

}
\author{
Bill Peterman
}
