% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgen_R_function.R
\name{PopGen_sim}
\alias{PopGen_sim}
\title{Internal PopGenReport function}
\usage{
PopGen_sim(
  sim_name = "output_",
  pts,
  sim_dir,
  resist_mat,
  n_ind = 25,
  n.offspring = 2,
  mig.rate = 0.2,
  disp.rate = 0.25,
  disp.max = 0.125,
  sex.ratio = 0.5,
  n.loci = 20,
  n.allels = 20,
  steps = 200,
  mut.rate = 5e-04
)
}
\arguments{
\item{sim_name}{Name for simulation results. Defaults to 'output'}

\item{pts}{Spatial points object}

\item{sim_dir}{Directory where simulation results will be written}

\item{resist_mat}{Resistance distance matrix}

\item{n_ind}{Default = 25; Number of individuals in each deme}

\item{n.offspring}{Default = 2; Poisson draw around ‘mean fecundity’}

\item{mig.rate}{Default = 0.2; Proportion of dispersing individuals that achieve the maximum dispersal distance}

\item{disp.rate}{Default = 0.25; Proportion of offspring that disperse}

\item{disp.max}{Default = 0.125; What percentage of the maximum pairwise distance can a dispersing individual traverse?}

\item{sex.ratio}{Default = 0.5}

\item{n.loci}{Default = 20}

\item{n.allels}{Default = 20}

\item{steps}{Number of iterations to run simulation}

\item{mut.rate}{Default = 0.0005}
}
\description{
Function to run PopGenReport from R
}
\author{
Bill Peterman
}
\keyword{internal}
