% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radishCV.R
\name{radish_cv}
\alias{radish_cv}
\title{radish cross validation}
\usage{
XXX
}
\arguments{
\item{pts}{SpatialPoints object of sampled demes}

\item{covariates}{RasterStack of raster layers for `radish` model}

\item{fmla}{Formula for model to be assessed. LHS should be the name of the distance matrix, RHS contains names of layers in `covariates`}

\item{model}{c('mlpe', 'wishart')}

\item{prop_train}{Proportion of sample points used for fitting a training model. The remainder of points will be used in assessment.}

\item{fit_full}{Should the full `radish` model with all sample demes be fit (Default = TRUE)}

\item{...}{Not used}
}
\value{
Named list containing the `radish` model fit to the training data, the loglikehood of the test data, and (optionally) the model fit to the complete data set.
}
\description{
Function to conduct cross validation of `radish` models
}
\examples{
## Not Run:
## ** TO BE COMPLETED ** ##  

## End (Not run)
}
\author{
Bill Peterman
}
