\name{radishDGS-package}
\alias{radishDGS-package}
\alias{radishDGS}
\docType{package}
\title{
\packageTitle{radishDGS}
}
\description{
\packageDescription{radishDGS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{radishDGS}
\packageIndices{radishDGS}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{radishDGS}

Maintainer: \packageMaintainer{radishDGS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
