% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cdpop.R
\name{map_cdpop}
\alias{map_cdpop}
\title{Map simulations}
\usage{
XXX
}
\arguments{
\item{params}{Object generated from `get_params` OR results from `cdpop_sim`}

\item{log}{(Default = TRUE) The the log transformed conductance surface will be plotted}

\item{...}{Not used}
}
\value{
Plot of conductance surface with sampled demes
}
\description{
Function to map simulated demes onto conductance surface
}
\examples{
## Not Run:
## map_cdpop(params,
             log = TRUE) ##  

## End (Not run)

}
\author{
Bill Peterman
}
